backend/requirements.txt
django
djangorestframework
pytest
psycopg2-binary==2.8.6
python-dotenv
djangorestframework-simplejwt==4.7.2
fpdf
django-cors-headers